+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+### 1.0.0-prerelease0010
+* Dom.OnTouchStart/Move/End/Cancel
+
 ### 1.0.0-prerelease0009
 * Sg PointerEvents get Button instead of int
 * Sg.Text no longer uses Aardvark.SceneGraph